refactoring asset_id.py to asset_base.py
